feat: skeleton state
